Changed the priority to long long Improved moves to the top of the transfer list

+++ include/megaapi.h
@@ -2711,7 +2711,7 @@ class MegaTransfer
          *
          * @return Priority of the transfer
          */
-        virtual float getPriority() const;
+        virtual unsigned long long getPriority() const;
 };
 
 /**
